Refactor some of the MultiParser code.
Create a non-member function for the code that jumps
to a checkpoint, and creates the appropriate GzipStreamReader
starting on a record boundary. This function also returns
the relevant context used by all of the other functions
that need this functionality (the number of records this stream
will yield, the number of bytes it will yield, and the
rank of the first record it will yield).